requirements.txt
appnope==0.1.4
asttokens==3.0.0
-e git+https://github.com/pradhanishan/bi-sales-simulator-spark.git@4dfc85d1bbbc395b6aa86c95dcc641435933f353#egg=bi.sales.simulator.spark
comm==0.2.3
debugpy==1.8.15
decorator==5.2.1
delta-spark==4.0.0
executing==2.2.0
Faker==37.4.2
importlib_metadata==8.7.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
Pygments==2.19.2
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyzmq==27.0.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
wcwidth==0.2.13
zipp==3.23.0
